New map selector.
A new map selector method added. This includes a new module called
inval, and a new menu called "Select map". When the user selects this
menu, she will be asked to tell if she wants a number greater or lesser
than the number offered. She can select using the up and down keys. This
method works like a binary search algorithm. My goal with this is to
avoid using the number keys. (Remember that we develop for PDA.)